Merges the scene manager from editor and app and moves to core, fixes incorrect rendering of beam
