+++ sscanss/core/scene/__init__.py
@@ -2,6 +2,6 @@ from .node import Node, BatchRenderNode, InstanceRenderNode
 from .entity import (SampleEntity, FiducialEntity, MeasurementPointEntity, MeasurementVectorEntity, InstrumentEntity,
                      PlaneEntity, BeamEntity)
 from .camera import Camera, world_to_screen, screen_to_world
-from .scene import Scene, validate_instrument_scene_size
+from .scene import Scene, SceneManager, validate_instrument_scene_size
 from .shader import VertexArray, GouraudShader, DefaultShader, Shader
 from .renderer import OpenGLRenderer
