+++ tests/test_ui.py
@@ -148,14 +148,14 @@ class TestMainWindow(unittest.TestCase):
 
         # render in transparent
         QTest.mouseClick(self.toolbar.widgetForAction(self.window.blend_render_action), Qt.LeftButton)
-        self.assertEqual(Scene.sample_render_mode, Node.RenderMode.Transparent)
+        self.assertEqual(self.window.scenes.sample_render_mode, Node.RenderMode.Transparent)
 
         self.keyinFiducials()
         self.keyinPoints()
 
         # render in wireframe
         QTest.mouseClick(self.toolbar.widgetForAction(self.window.line_render_action), Qt.LeftButton)
-        self.assertEqual(Scene.sample_render_mode, Node.RenderMode.Wireframe)
+        self.assertEqual(self.window.scenes.sample_render_mode, Node.RenderMode.Wireframe)
 
         self.insertVectors()
         self.jawControl()
